feat(wgpu): Implement multiplication operation
This commit implements the multiplication operation in the wgpu module. The `OpMul` struct is added to the `op` module and the `mul` method is added to the `GPUTensor` struct. The `OpMul` struct implements the `Op` trait and is responsible for setting up the shader and calculating the workgroups for the multiplication operation. The `mul` method in the `GPUTensor` struct creates a new variable using the `VarType::Mul` enum variant and performs the multiplication operation. Lines of code should not exceed 74 characters.